Reglage PB nom champ base de donnees + debut visualisation programme d'un utilisateur

+++ Projet-JEE/src/main/java/com/example/Projet_JEE/entity/Programme_therapeutique_activite_Id.java
@@ -5,6 +5,6 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public class Programme_therapeutique_activite_Id {
 
-    private int idProgrammeTherapeutique;
-    private int idActivite;
+    private Long idProgrammeTherapeutique;
+    private Long idActivite;
 }
